refactor: move parsing method to Lotto class
